接入360

+++ tamguo-crawler/src/main/java/com/tamguo/model/CourseEntity.java
@@ -4,9 +4,6 @@ import java.io.Serializable;
 import com.baomidou.mybatisplus.annotations.TableName;
 import com.tamguo.config.dao.SuperEntity;
 
-import java.math.BigInteger;
-
-
 /**
  * The persistent class for the tiku_course database table.
  * 
@@ -19,9 +16,9 @@ public class CourseEntity extends SuperEntity<CourseEntity> implements Serializa
 
 	private String subjectId;
 	
-	private BigInteger pointNum;
+	private Integer pointNum;
 	
-	private BigInteger questionNum;
+	private Integer questionNum;
 
 	private Integer orders;
 	
@@ -50,19 +47,19 @@ public class CourseEntity extends SuperEntity<CourseEntity> implements Serializa
 		this.subjectId = subjectId;
 	}
 
-	public BigInteger getQuestionNum() {
+	public Integer getQuestionNum() {
 		return questionNum;
 	}
 
-	public void setQuestionNum(BigInteger questionNum) {
+	public void setQuestionNum(Integer questionNum) {
 		this.questionNum = questionNum;
 	}
 
-	public BigInteger getPointNum() {
+	public Integer getPointNum() {
 		return pointNum;
 	}
 
-	public void setPointNum(BigInteger pointNum) {
+	public void setPointNum(Integer pointNum) {
 		this.pointNum = pointNum;
 	}
